Add files via upload
Included in this commit is the code used to render our streamlit application. A lot is still not general, but its best to share this now. The UI team is working on making it more general as well as cleaning it up a bit. More changes will come soon. 

A requirements.txt file was also created using a python virtual environment  and running the pip freeze command. That file is also uploaded. The file seems to have grabbed a lot more libraries than were specified, but those might also be additional dependencies.